backend/phase1_intent_detection_routing/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.0.0
python-dotenv>=1.0.0
groq>=0.9.0

# Phase 2 dependencies
google-generativeai>=0.8.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.150.0
python-multipart>=0.0.12

# RAG dependencies
chromadb>=0.5.0
PyPDF2>=3.0.0
python-docx>=1.1.0
openpyxl>=3.1.0

# MCP dependencies
mcp>=1.0.0
nest-asyncio>=1.5.0

# Phase 6 dependencies
websockets>=12.0
sarvamai>=0.1.0
